Iteration 2.1
Scheduler - updated to specify items being manipulated in the queues as pickups or drop offs
ElevatorQueue - keeps pickups and drop offs seperate so the scheduler can better inform the floor subsystem.
ElevatorSubSystem - ArrivalSensor has been in the elevator subsystem spec for a while now, finally got around to moving it in. UML will need updating.
FloorSubsystem - removed the arrival sensors and reworked how the direction lamps are used and updated based on the schedulers intention.