Doors Stuck Open
This change makes it so doors are only close when an up message is sent, a down message is sent, or the elevator lets off it's last passenger.
In addition a pending soft fault will only enter an active state when the doors are left open at the end of an instruction.
The reason for this change is to more accurately model the 'doors stuck open' fault as the ui will need to display this info.